move projects API to the right place
